bewise/requirements.txt
pytz==2022.1
requests==2.28.1
setuptools==63.4.2
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
psycopg2-binary==2.9.3
reportlab==3.6.11
python-dotenv==0.20.0
django-environ==0.9.0